Move DoubleBooleanNegation to Refactor category
Also tries to improve the explanation
